requirements.txt
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dnspython==2.8.0
fastapi==0.118.0
ffmpeg-python==0.2.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
h11==0.16.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.6
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.6.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.48.0
streamlit==1.50.0
streamlit-webrtc==0.63.11
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
webrtcvad==2.0.10
websockets==15.0.1
opencv-python==4.12.0.88